+++ jest.config.js
@@ -5,7 +5,7 @@ module.exports = {
     moduleNameMapper: {
         '^@/(.*)$': '<rootDir>/src/$1',
     },
-    testMatch: ['**/src/util/**/*.test.(ts|tsx)'],
+    testMatch: ['**/src/util/**/*.test.(ts|tsx)','**/src/commands/fts/SearchWorkbench/test/*.test.ts'],
     transform: {
         '^.+\\.(ts|tsx)$': 'ts-jest',
     },