+++ src/extension.ts
@@ -52,7 +52,7 @@ import { WorkbenchWebviewProvider } from "./workbench/workbenchWebviewProvider";
 import { fetchClusterOverview } from "./pages/overviewCluster/overviewCluster";
 import DependenciesDownloader from "./handlers/handleCLIDownloader";
 import { sqlppFormatter } from "./commands/sqlpp/sqlppFormatter";
-import { fetchQueryContext } from "./pages/queryContext/queryContext";
+import { fetchQueryContext, fetchSearchContext } from "./pages/queryContext/queryContext";
 import { fetchFavoriteQueries } from "./pages/FavoriteQueries/FavoriteQueries";
 import { markFavoriteQuery } from "./commands/favoriteQueries/markFavoriteQuery";
 import { QueryHistoryTreeProvider } from "./tree/QueryHistoryTreeProvider";
@@ -78,6 +78,14 @@ import { kvTypeFilterDocuments } from "./commands/documents/documentFilters/kvTy
 import { fetchNamedParameters } from "./pages/namedParameters/namedParameters";
 import { sqlppComlpletions, sqlppNamedParametersCompletions, sqlppSchemaComlpletions } from "./commands/sqlpp/sqlppCompletions";
 import { dynamodbMigrate } from "./pages/Tools/DynamoDbMigrate/dynamoDbMigrate";
+import { SearchWorkbench } from "./commands/fts/SearchWorkbench/searchWorkbench";
+import SearchIndexNode from "./model/SearchIndexNode";
+import { openSearchIndex } from "./commands/fts/SearchWorkbench/openSearchIndex";
+import { handleSearchContextStatusbar } from "./handlers/handleSearchQueryContextStatusBar";
+import { validateDocument } from "./commands/fts/SearchWorkbench/validators/validationUtil";
+import { AutocompleteVisitor } from "./commands/fts/SearchWorkbench/contributor/autoCompleteVisitor";
+import { CbsJsonHoverProvider } from "./commands/fts/SearchWorkbench/documentation/documentationProvider";
+import { deleteIndex } from "./util/ftsIndexUtils";
 
 export function activate(context: vscode.ExtensionContext) {
   Global.setState(context.globalState);
@@ -94,6 +102,64 @@ export function activate(context: vscode.ExtensionContext) {
 
   const uriToCasMap = new Map<string, string>();
   const workbench = new QueryWorkbench();
+  const searchWorkbench = new SearchWorkbench();
+
+  let currentSearchIndexNode: SearchIndexNode;
+  let currentSearchWorkbench: SearchWorkbench;
+
+  let diagnosticCollection = vscode.languages.createDiagnosticCollection('jsonValidation');
+
+  diagnosticCollection = vscode.languages.createDiagnosticCollection('jsonValidation');
+  context.subscriptions.push(diagnosticCollection);
+
+context.subscriptions.push(vscode.workspace.onDidChangeTextDocument(event => {
+    if (event.document === vscode.window.activeTextEditor?.document && event.document.languageId == "json" && vscode.window.activeTextEditor?.document.fileName.endsWith(".cbs.json")) {
+        validateDocument(event.document, diagnosticCollection);
+    }
+}));
+
+const hoverProvider = new CbsJsonHoverProvider(context);
+context.subscriptions.push(
+    vscode.languages.registerHoverProvider({ language: 'json', pattern: '**/*.cbs.json' }, hoverProvider)
+);
+
+const provider = vscode.languages.registerCompletionItemProvider(
+  { language: 'json', pattern: '**/*.cbs.json' },
+  {
+      async provideCompletionItems(document, position, token, context) {
+          const autoComplete = new AutocompleteVisitor();
+          const suggestions = await autoComplete.getAutoCompleteContributor(document, position, currentSearchWorkbench, cacheService);
+          if (suggestions.length === 0) {
+            return [new vscode.CompletionItem('', vscode.CompletionItemKind.Text)].map(item => {
+                item.sortText = '\u0000';
+                item.preselect = true;
+                item.keepWhitespace = true;
+                item.insertText = '';
+                item.range = new vscode.Range(position, position);
+                return item;
+            });
+        }
+          return suggestions;
+      }
+  },
+  '\"' 
+);
+
+
+
+context.subscriptions.push(provider);
+
+const disposable = vscode.window.onDidChangeTextEditorSelection(async (e) => {
+  if (e.kind === vscode.TextEditorSelectionChangeKind.Command) {
+      const activeEditor = vscode.window.activeTextEditor;
+      if (activeEditor && activeEditor.document.fileName.endsWith('.cbs.json')) {
+          await vscode.commands.executeCommand('editor.action.formatDocument');
+      }
+  }
+});
+
+context.subscriptions.push(disposable);
+
 
   const subscriptions = context.subscriptions;
   const cacheService = new CacheService();
@@ -187,7 +253,9 @@ export function activate(context: vscode.ExtensionContext) {
       if (
         editor &&
         editor.document.languageId === "json" &&
-        editor.document.uri.scheme === "couchbase"
+        editor.document.uri.scheme === "couchbase" &&
+        !editor.document.uri.path.endsWith("cbs.json") &&
+        !editor.document.uri.path.includes("/Search/")
       ) {
         await handleActiveEditorChange(editor, uriToCasMap, memFs);
       }
@@ -199,7 +267,9 @@ export function activate(context: vscode.ExtensionContext) {
       async (document: vscode.TextDocument) => {
         if (
           document && document.languageId === "json" &&
-          document.uri.scheme === "couchbase"
+          document.uri.scheme === "couchbase" && 
+          !document.uri.path.endsWith("cbs.json") &&
+          !document.uri.path.includes("/Search/")
         ) {
           await handleOnSaveTextDocument(document, uriToCasMap, memFs, cacheService);
           clusterConnectionTreeProvider.refresh();
@@ -289,6 +359,25 @@ export function activate(context: vscode.ExtensionContext) {
       Commands.openDocument,
       async (documentNode: DocumentNode) => {
         await openDocument(documentNode, clusterConnectionTreeProvider, uriToCasMap, memFs);
+      }
+    )
+  );
+
+  subscriptions.push(
+    vscode.commands.registerCommand(
+      Commands.openSearchIndex,
+      async (searchIndexNode: SearchIndexNode) => {
+        await openSearchIndex(searchIndexNode, clusterConnectionTreeProvider, uriToCasMap, memFs);
+      }
+    )
+  );
+
+  subscriptions.push(
+    vscode.commands.registerCommand(
+      Commands.deleteSearchIndex,
+      async (searchIndexNode: SearchIndexNode) => {
+        await deleteIndex(searchIndexNode);
+        clusterConnectionTreeProvider.refresh();
       }
     )
   );
@@ -522,6 +611,10 @@ export function activate(context: vscode.ExtensionContext) {
     )
   );
 
+  // Initialize the global status bar item which will be used for query and search query context
+  let globalStatusBarItem = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, 1000);
+  globalStatusBarItem.hide(); 
+
   vscode.languages.registerDocumentFormattingEditProvider('SQL++', {
     provideDocumentFormattingEdits(document: vscode.TextDocument): vscode.TextEdit[] {
       return sqlppFormatter(document);
@@ -607,19 +700,19 @@ export function activate(context: vscode.ExtensionContext) {
 
   subscriptions.push(
     vscode.commands.registerCommand(Commands.queryContext, () => {
-      fetchQueryContext(workbench, context);
+      fetchQueryContext(workbench, context, globalStatusBarItem);
     })
   );
 
   // subscription to make sure query context status bar is only visible on sqlpp files
   subscriptions.push(
     vscode.window.onDidChangeActiveTextEditor(async (editor) => {
-      await handleQueryContextStatusbar(editor, workbench);
+      await handleQueryContextStatusbar(editor, workbench, globalStatusBarItem);
     })
   );
-  // Handle initial view of context status bar
+  // // Handle initial view of context status bar
   const activeEditor = vscode.window.activeTextEditor;
-  handleQueryContextStatusbar(activeEditor, workbench);
+  handleQueryContextStatusbar(activeEditor, workbench, globalStatusBarItem);
 
   subscriptions.push(
     vscode.commands.registerCommand(
@@ -643,7 +736,7 @@ export function activate(context: vscode.ExtensionContext) {
     vscode.commands.registerCommand(
       Commands.showNamedParameters,
       () => {
-         fetchNamedParameters();
+        fetchNamedParameters();
       }
     )
   );
@@ -702,6 +795,39 @@ export function activate(context: vscode.ExtensionContext) {
     })
   );
 
+  const searchContextCommand = vscode.commands.registerCommand(Commands.searchContext, () => {
+    fetchSearchContext(currentSearchIndexNode, currentSearchWorkbench, context, globalStatusBarItem);
+  });
+  context.subscriptions.push(searchContextCommand);
+
+
+
+  const openSearchWorkbenchCommand = vscode.commands.registerCommand(Commands.openSearchWorkbench, async (searchIndexNode: SearchIndexNode) => {
+    const connection = Memory.state.get<IConnection>(Constants.ACTIVE_CONNECTION);
+    if (!connection) {
+      vscode.window.showErrorMessage("No active connection available.");
+      return;
+    }
+
+    currentSearchIndexNode = searchIndexNode;
+    currentSearchWorkbench = searchWorkbench;
+
+    searchWorkbench.openSearchWorkbench(searchIndexNode, memFs);
+
+
+  });
+  context.subscriptions.push(openSearchWorkbenchCommand);
+
+
+  subscriptions.push(
+    vscode.window.onDidChangeActiveTextEditor(async (editor) => {
+      if (editor && editor.document.languageId === "json" && editor.document.fileName.endsWith(".cbs.json")) {
+        await handleSearchContextStatusbar(editor, currentSearchIndexNode, searchWorkbench, globalStatusBarItem);
+      }
+    })
+  );
+
+
   context.subscriptions.push(
     vscode.workspace.registerNotebookSerializer(
       Constants.notebookType, new QueryContentSerializer(), { transientOutputs: true }
@@ -732,6 +858,30 @@ export function activate(context: vscode.ExtensionContext) {
     "vscode-couchbase.runButtonEnabled",
     true
   ); // Required to enable run query button at the start
+
+  context.subscriptions.push(
+    vscode.commands.registerCommand(Commands.runSearchQuery, async (searchIndexNode: SearchIndexNode) => {
+      vscode.commands.executeCommand(
+        "setContext",
+        "vscode-couchbase.runSearchButtonEnabled",
+        undefined
+      );
+      await searchWorkbench.runCouchbaseSearchQuery(
+        workbenchWebviewProvider
+      );
+      vscode.commands.executeCommand(
+        "setContext",
+        "vscode-couchbase.runSearchButtonEnabled",
+        true
+      );
+    })
+  );
+  vscode.commands.executeCommand(
+    "setContext",
+    "vscode-couchbase.runSearchButtonEnabled",
+    true
+  ); // Required to enable run search query button at the start
+
 
   context.subscriptions.push(
     vscode.commands.registerCommand(